fix: jenkins connection use helper entity
